+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Change Log
 
 ### New Features
 
+- Add a new scope “syntax style” to the snippet feature to perform snippets only in a specific sytnax style.
 - Add a new variable “selection” to the snippet feature to place the current selected text to the inserted text.